# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface InsertContact {
  name: string;
  email: string;
  company?: string | null;
  message: string;
}

export interface InsertProject {
  name: string;
  email: string;
  company?: string | null;
  projectType: string;
  budget: string;
  timeline: string;
  description: string;
  features: string[];
}

export interface SubmitSuccess {
  success: boolean;
  id: string;
}

export interface ErrorResponse {
  error: string;
}