# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact form submissions
  - name: project
    description: Project inquiry submissions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertContact"
      responses:
        "200":
          description: Contact submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitSuccess"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /project:
    post:
      operationId: submitProject
      tags: [project]
      summary: Submit project inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertProject"
      responses:
        "200":
          description: Project submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitSuccess"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    InsertContact:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        company:
          type: string
          nullable: true
        message:
          type: string
      required:
        - name
        - email
        - message
    InsertProject:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        company:
          type: string
          nullable: true
        projectType:
          type: string
        budget:
          type: string
        timeline:
          type: string
        description:
          type: string
        features:
          type: array
          items:
            type: string
      required:
        - name
        - email
        - projectType
        - budget
        - timeline
        - description
        - features
    SubmitSuccess:
      type: object
      properties:
        success:
          type: boolean
        id:
          type: string
      required:
        - success
        - id
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error