# lib/api-zod/src/generated/types/insertContact.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InsertContact {
  name: string;
  email: string;
  company?: string | null;
  message: string;
}