# lib/api-zod/src/generated/types/insertProject.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InsertProject {
  name: string;
  email: string;
  company?: string | null;
  projectType: string;
  budget: string;
  timeline: string;
  description: string;
  features: string[];
}